[doc,man]: Fix botched description of "unitwidth".
I had made a hash of this exposition, replacing a different hash that
came before, and which even Kernighan did not illuminate clearly in AT&T
troff documentation--at least not for me.

The most important fact--either omitted, contradicted, or woefully
underemphasized in earlier device-independent troff documentation--is
that the "unit width" is _not_ "basic units per point".  It _could_ be,
but need not be, otherwise the unit width would have to be coupled to
the device resolution, which is in basic units per _inch_, and a point
is 1/72 inch.

But no such constraint exists, and none is necessary.  I explain why.

Everybody should learn at least unit analysis, and ideally its more
abstract cousin, dimensional analysis.  They prevent errors like these.