groff_man*(7): Migrate to `\[xx]` escape form.
...from `\(xx` in page source and illustrations of escape sequence
syntax.

A few years of experience corresponding with man page maintainers in
other projects has persuaded me that the proportion thereof who are
concerned with page portability to AT&T troff is low.  The only
surviving instances appear to be Plan 9 troff and Solaris 10 troff.
(I don't think DWB 3.3 troff is anyone's "daily driver", but it's
valuable for research and testing purposes.)

Nevertheless in a subsequent commit I'll add a Q&A item to
groff_man_style(7)'s "Notes" section covering this topic.