groff_mdoc(7): Drop garbage spaces from table.
They were throwing off some cell alignments in PDF output with one
viewer[*] (the one built in to Dropbox, which is a frustrating and
disappointing PDF reader in multiple respects).

That reader could very well be buggy, but this change is the right thing
to do anyway, and relieves us of having to use the GNU tbl "nospaces"
region option extension.

[*] I did not observe problems with xpdf, evince, or okular.