doc/groff.texi.in: Clarify non-deprecation of \".
A contributor to Vim interpreted our Texinfo manual's statement "GNU
troff introduced this extension [`\#`] to avoid the problems described
above" as establishing a preference for `\#` over `\"`--but this is not
so, as all groff maintainers from James Clark forward have used `\"`
freely when composing macro files and documents.  Elaborate the example
and present a pitfall of using `\#` robotically.

See <https://github.com/vim/vim/pull/17516>.