doc/*,man/*: Update sentence-ending punct advice.
It's not necessary to "guard" a non-sentential [.?!] from a subsequent
tab character; the formatter represents tab stop advancement by a
horizontal motion, which is "not space", so it undergoes neither
ordinary adjustment not inter-sentence space supplementation.

When filling is enabled, a newline after a tab becomes an adjustable
word space as usual.

Here's an illustration using leaders, which work the same way as tabs,
but are more visible in output.

$ cat EXPERIMENTS/sentence-ending-with-leader.roff | tr '\001' '@'
Here's a sentence.
A second sentence.
.br
A sentence followed by a tab?@Hello!
.br
A sentence followed by a tab?@
Hello!
.br
A sentence followed by a tab?\&@
Hello!
$ nroff EXPERIMENTS/sentence-ending-with-leader.roff | cat -s
Here’s a sentence.  A second sentence.
A sentence followed by a tab?...Hello!
A sentence followed by a tab?... Hello!
A sentence followed by a tab?... Hello!

Also add additional relevant concept index entries for Texinfo.

Also align input line breaks between our Texinfo manual and roff(7).