ChangeLog: Tweak format of recent entry.
https://www.gnu.org/prep/standards/standards.html#Style-of-Change-Logs

I tend to use a sort of module name in brackets for the first line
of a Git commit message and sometimes re-use that in the ChangeLog entry
itself where it seems useful to summarize a change.  This "module name"
serves to abbreviate lengthy source file names to buy space for
description in Git's 50-character (soft) limit on a commit message's
first line, and (usually) corresponds to the "Category" in groff's bug
tracker on Savannah.

I also try to leave out the "index.php" part of Savannah ticket URLs
because (1) it's unnecessary, and (2) I hope that, someday, Savannah
will migrate to a content management system that isn't based on the
awful PHP language.