doc/*,man/*: Reform terminology.
"additional inter-sentence space" -> "supplemental inter-sentence space"

The latter is the phraseology I have semi-consciously drifted to when
coaching users in email; in this case, I regard that as the voice of my
muse.  Further, it's a better mnemonic for the `.sss` register name we
use to store the corresponding amount thereof in the environment.