src/roff/troff/node.cpp: Trivially refactor.
`unsigned char` is a _numeric_ type, not a character type (properly
considered).  Compare values of this type to unsigned numeric literals,
not to character literals that are to be used with `char`, which is of
undefined signedness.  Reorder equality comparisons to avoid inadvertent
lvalue assignment.

* src/roff/troff/node.cpp (troff_output_file::put_char_width)
  (troff_output_file::put_char, glyph_node::get_hyphenation_type)
  (node::asciify, composite_node::dump_properties, make_glyph_node): Do
  it.

  (charinfo_node::dump_properties): Perform explicit comparison to
  unsigned numeric literal instead of letting its value pun down to a
  Boolean.

(It might be worth pointing out that `unsigned char` _is_ GNU troff's
internal data type for representing character codes.  It is, therefore,
the type we'll want to migrate to `int32_t` for graceful internal
handling of the Unicode character set.  See Savannah #40720.)