contrib/gpinyin/ChangeLog: Revise copyright info.
Add a copyright notice for myself.  I've never executed copyright
assignment paperwork with the FSF, so my "legally significant"[1]
changes to this file in 2021-2022 can't be under the FSF's copyright.

Drop ersatz '(C)' copyright symbol.  Software developers have long
labored under the no-longer-correct misconception that omitting a
copyright symbol from one's notice was a fatal defect that effectively
placed the work in the public domain.  That stopped being true as of 1
March 1989.[2]  Further, prior to guidance issued by the U.S. Copyright
Office in the decades since, the use of "(C)" as a substitute for a
copyright sign _may not have sufficed_ to prevent the copyright notice
from being regarded as defective.  The Copyright Office, then and now,
prefers the abbreviation "copr." when © is typographically
unavailable.[2]  Nowadays, its advice is that "c" (note lowercase) is an
"acceptable variant", that _may_ retain the efficacy of the copyright
notice.  The word "copyright", spelled out in full, also suffices per
that resource, and is already present in this notice.

Use fuller form of traditional GNU permissive documentation license,
including the phrase "in any medium without royalty".  If those
qualifiers were assumed to not be in place previously, then this
document would not have been Free Software, and I assume that was no
one's intention.

Drop spurious line at end of log that records no change.  Maybe it was
there for convenience of copy-and-paste operations.

Drop "written by" annotation--it is redundant with the chronologically
first log entry, which asserts "First version 0.9.0" as Bernd's work,
and it does not accurately characterize later work done by others; this
is why change log entries themselves carry authorship credits.

[1] "A change of just a few lines (less than 15 or so) is not legally
    significant for copyright."  Conversely, >= 15 lines would be.

    https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html

[2] https://www.copyright.gov/circs/circ03.pdf