tmac/an.tmac: Revise copyright notice.
Add a copyright notice for myself.  I've never executed copyright
assignment paperwork with the FSF, so my "legally significant"[1]
changes to this file in 2020-2025[2] can't be under the FSF's copyright.

(I retain an FSF copyright notice corresponding to all changes not
authored by me [even where in a given year no such individual change
meets the aforementioned threshold of "legal significance"--maybe I
shouldn't in this case], because I don't know who _has_ filed copyright
assignment paperwork with the FSF; I know only that I haven't.)

Drop ersatz '(C)' copyright symbol.  Software developers have long
labored under the no-longer-correct misconception that omitting a
copyright symbol from one's notice was a fatal defect that effectively
placed the work in the public domain.  That stopped being true as of 1
March 1989.[3]  Further, prior to guidance issued by the U.S. Copyright
Office in the decades since, the use of "(C)" as a substitute for a
copyright sign _may not have sufficed_ to prevent the copyright notice
from being regarded as defective.  The Copyright Office, then and now,
prefers the abbreviation "copr." when © is typographically
unavailable.[3]  Nowadays, its advice is that "c" (note lowercase) is an
"acceptable variant", that _may_ retain the efficacy of the copyright
notice.  The word "copyright", spelled out in full, also suffices per
that resource, and is already present in this notice.

[1] "A change of just a few lines (less than 15 or so) is not legally
    significant for copyright."  Conversely, >= 15 lines would be.

    https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html

[2] Here is a qualifying commit from each calendar year in the range,
    produced with the aid of `git --log --stat --follow`.

commit 54fb0303cfd59873dd9dc8aab3a47f1e7eff1341
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Tue Jul 15 23:15:56 2025 -0500

    [man]: Fix Savannah #67363 (2/4).

 tmac/an.tmac | 56 ++++++++++++++++++++++++++++++++++++++++++++++++++++----

commit d1a148f95cbf1ecd21c897e8882817e9f7ee371d
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Sat Jan 27 16:54:10 2024 -0600

    tmac/an.tmac: Refactor (an*bookmark*DEVICE).

 tmac/an.tmac | 17 ++++++++++++++++-

commit 5485e3b7f1be2956e2067d572c6ce8fd5c036fc1
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Fri Aug 4 09:28:55 2023 -0500

    [man]: Deprecate `SB` macro.

 tmac/an.tmac | 19 +++++++++++++++++--

commit 478494e32112aabaa04f8314b5305c7bf6ea1689
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Mon Dec 12 10:31:19 2022 -0600

    [man, mdoc]: Fix Savannah #63500.

 tmac/an.tmac | 39 ++++++++++++++++++++++++++++++++++-----

commit 8578254c09b9fd66cad59353caae479240037cbf
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Tue Oct 26 16:15:32 2021 +1100

    [man]: Abbreviate the inside footer if necessary.

 tmac/an.tmac | 66 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++---

[between the foregoing commit and the next, the file was renamed]

commit 25d2a812faf52e4d952769920ccfbe5e61cb1b43
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Sat Nov 14 13:25:26 2020 +1100

    tmac/an-old.tmac: Add style checks.

 tmac/an-old.tmac | 23 +++++++++++++++++++++++

[3] https://www.copyright.gov/circs/circ03.pdf