tmac/doc.tmac: Revise copyright notice.
Add a copyright notice for myself.  I've never executed copyright
assignment paperwork with the FSF, so my "legally significant"[1]
changes to this file in 2021-2024[2] can't be under the FSF's copyright.

(I retain an FSF copyright notice corresponding to all changes not
authored by me [even where in a given year no such individual change
meets the aforementioned threshold of "legal significance"--maybe I
shouldn't in this case], because I don't know who _has_ filed copyright
assignment paperwork with the FSF; I know only that I haven't.)

An analysis predicated on "a net increase of at least 15 lines in one
commit" results in a ratcheting back of the dates in the notice.  I'm
the only person to have contributed to this file since Ingo Schwarze in
2019, and by this metric, I made no single "legally significant" change
this calendar year.  I can apply the technique discussed in
our "HACKING" file to make further revisions to the date--but not for
this commit.

Drop ersatz '(C)' copyright symbol.  Software developers have long
labored under the no-longer-correct misconception that omitting a
copyright symbol from one's notice was a fatal defect that effectively
placed the work in the public domain.  That stopped being true as of 1
March 1989.[3]  Further, prior to guidance issued by the U.S. Copyright
Office in the decades since, the use of "(C)" as a substitute for a
copyright sign _may not have sufficed_ to prevent the copyright notice
from being regarded as defective.  The Copyright Office, then and now,
prefers the abbreviation "copr." when © is typographically
unavailable.[3]  Nowadays, its advice is that "c" (note lowercase) is an
"acceptable variant", that _may_ retain the efficacy of the copyright
notice.  The word "copyright", spelled out in full, also suffices per
that resource, and is already present in this notice.

[1] "A change of just a few lines (less than 15 or so) is not legally
    significant for copyright."  Conversely, >= 15 lines would be.

    https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html

[2] Here is a qualifying commit from applicable calendar years in the
    range, produced with the aid of `git --log --stat --follow`.

commit 3f7c24349d420789f2a465ff39a337d25f21caa6
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Tue Mar 19 14:51:00 2024 -0500

    [mdoc]: Set literal displays in Courier family.

 tmac/doc.tmac | 31 +++++++++++++++++++++++++++----

commit 78922cf9f39f6ebcd3a80aaab9c3a2931b8b0ad4
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Tue Dec 27 02:46:19 2022 -0600

    [mdoc]: Support `HF` string.

 tmac/doc.tmac | 19 ++++++++++++++++++-

commit ee41d36200ffc989eac09e65db7537bf59dd1c62
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Sun Jul 4 23:56:18 2021 +1000

    Skip the stripper, part 3/3 (mdoc). [part 3c]

 tmac/doc.tmac | 291 ++++++++++++++++++++++++++++++++++++++++++++++++--

[3] https://www.copyright.gov/circs/circ03.pdf