tmac/groff_man.7.man.in: Revise copyright notice.
Add a copyright notice for myself.  I've never executed copyright
assignment paperwork with the FSF, so my "legally significant"[1]
changes to this file in 2017-2025[2] can't be under the FSF's copyright.

(I retain an FSF copyright notice corresponding to all changes not
authored by me, because I don't know who _has_ filed copyright
assignment paperwork with the FSF; I know only that I haven't.)

Drop ersatz '(C)' copyright symbol.  Software developers have long
labored under the no-longer-correct misconception that omitting a
copyright symbol from one's notice was a fatal defect that effectively
placed the work in the public domain.  That stopped being true as of 1
March 1989.[3]  Further, prior to guidance issued by the U.S. Copyright
Office in the decades since, the use of "(C)" as a substitute for a
copyright sign _may not have sufficed_ to prevent the copyright notice
from being regarded as defective.  The Copyright Office, then and now,
prefers the abbreviation "copr." when © is typographically
unavailable.[3]  Nowadays, its advice is that "c" (note lowercase) is an
"acceptable variant", that _may_ retain the efficacy of the copyright
notice.  The word "copyright", spelled out in full, also suffices per
that resource, and is already present in this notice.

[1] "A change of just a few lines (less than 15 or so) is not legally
significant for copyright."  Conversely, >= 15 lines would be.

https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html

[2] For evidence of such contributions, one is spoiled for choice.
    Consider these commits from each calendar year.

2025: fa8b882f032d385efe295d89477b654234c58f8a
2024: 74c1595c4184c610984e429a04be76352b44869f
2023: 23aa71d14a96c635cf3a64d78e12924cebd07e64
2022: 54119b8931dfe6668cd06716bdd79a75dfe74549
2021: 0a111a916ee344d5046e244dcc95b4d3f49c352d
2020: a7c3446ca1d2e4ce516ec89fa154bdf8cec123e9
2019: 9503b794e821ef1cf6f705b25dc7abbadb920ad2
2018: 4c8775e39cc13c3be71448536d4fc5fe76319efc
2017: e59d0b4b9810539e359d0037e8e1751b68f76167

[3] https://www.copyright.gov/circs/circ03.pdf