tmac/an-ext.tmac: Revise copyright notice.
Add a copyright notice for myself.  I've never executed copyright
assignment paperwork with the FSF, so my "legally significant"[1]
changes to this file in 2021-2022 and 2024[2] can't be under the FSF's
copyright.

Drop ersatz '(C)' copyright symbol.  Software developers have long
labored under the no-longer-correct misconception that omitting a
copyright symbol from one's notice was a fatal defect that effectively
placed the work in the public domain.  That stopped being true as of 1
March 1989.[3]  Further, prior to guidance issued by the U.S. Copyright
Office in the decades since, the use of "(C)" as a substitute for a
copyright sign _may not have sufficed_ to prevent the copyright notice
from being regarded as defective.  The Copyright Office, then and now,
prefers the abbreviation "copr." when © is typographically
unavailable.[ibid.]  Nowadays, its advice is that "c" (note lowercase)
is an "acceptable variant", that _may_ retain the efficacy of the
copyright notice.  The word "copyright", spelled out in full, also
suffices per that resource, and is already present in this notice.

[1] "A change of just a few lines (less than 15 or so) is not legally
    significant for copyright."  Conversely, >= 15 lines would be.

    https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html

[2] Here are commits from each calendar year in that (disjunct) range
    that (cumulatively) manfiest "legal significance", produced with the
    aid of `git --log --stat`.

2021
====

commit 5c250cc6e79c998135b5c30582d28d9565693846
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Fri Oct 1 21:43:31 2021 +1000

    [man]: Use OSC 8 hyperlinks.

 tmac/an-ext.tmac | 50 ++++++++++++++++++++++++++++++++++++------------
 1 file changed, 38 insertions(+), 12 deletions(-)

2022
====

commit 096c2f056746f4fd0de3ed7e060c58e5e8d6afb5
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Wed Feb 16 16:03:59 2022 +1100

    tmac/an-ext.tmac: Slightly refactor hyphenation.

 tmac/an-ext.tmac | 18 ++++++++++++------
 1 file changed, 12 insertions(+), 6 deletions(-)

commit 21e79bcdf2780449f6f4ce094ce03b1bd4a1e3b4
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Tue Jun 21 06:44:15 2022 -0500

    [man]: Fix problem with hyphenation of link text.

 tmac/an-ext.tmac | 12 ++++++++++++
 1 file changed, 12 insertions(+)

2024
====

commit a08908f982c1b1ff8489aa0f857e6b10b76f05d6
Author: G. Branden Robinson <g.branden.robinson@gmail.com>
Date:   Fri May 3 14:53:45 2024 -0500

    [man]: Revise and expand `SY` and `YS` behavior.

 tmac/an-ext.tmac | 62 +++++++++++++++++++++++++++++++++---------------
 1 file changed, 43 insertions(+), 19 deletions(-)

[3] https://www.copyright.gov/circs/circ03.pdf