NEWS: Revise copyright notice.
Add copyright notices for myself and Ingo Schwarze.  Ingo and I have
never executed copyright assignment paperwork with the FSF, so our
"legally significant"[1] changes to this file in 2018-2025[2] can't be
under the FSF's copyright.

(I retain an FSF copyright notice corresponding to all changes not
authored by me or Ingo--and which are above the "legal significance"
threshold--because I don't know who _has_ filed copyright assignment
paperwork with the FSF; I know only that Ingo and I haven't.)

Drop ersatz '(C)' copyright symbol.  Software developers have long
labored under the no-longer-correct misconception that omitting a
copyright symbol from one's notice was a fatal defect that effectively
placed the work in the public domain.  That stopped being true as of 1
March 1989.[3]  Further, prior to guidance issued by the U.S. Copyright
Office in the decades since, the use of "(C)" as a substitute for a
copyright sign _may not have sufficed_ to prevent the copyright notice
from being regarded as defective.  The Copyright Office, then and now,
prefers the abbreviation "copr." when © is typographically
unavailable.[ibid.]  Nowadays, its advice is that "c" (note lowercase)
is an "acceptable variant", that _may_ retain the efficacy of the
copyright notice.  The word "copyright", spelled out in full, also
suffices per that resource, and is already present in this notice.

[1] "A change of just a few lines (less than 15 or so) is not legally
significant for copyright."  Conversely, >= 15 lines would be.

https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html

[2] Regarding Ingo, see commit 3805d2a0e4, 12 April 2022.

    Regarding evidence of my contributions, one is spoiled for choice.
    Consider these commits from each calendar year.

2025: db04b8d2cd90e7aa49f65c71d7bb023f3922fb50
2024: 1cfb27606c47adac2720e664c7fd3f155bdf280d
2023: 1058ed9c474c9e0689c5c55670175236f8a3bfc5
2022: 915a878038236769eb072f728389352c1da88719
2021: d2def30b3e87c2513f02bc50be54dee700fa7125
2020: 0b9aaca060117edfc65af8264fa791ed7e851633
2019: 2e2d302aa7bce88bad9b05f24db22b4727957f69
2018: 4c8775e39cc13c3be71448536d4fc5fe76319efc
      3fa04f3e60bebfdcf968f573e6b46e6236a55789
      93f1a2060c11daf36d9b7099ef2720d644937451
      b60f6b6d0c0a9655d10d32ee6688ac0f45fc3006

[3] https://www.copyright.gov/circs/circ03.pdf