doc/ms.ms.in: Revert mistaken change.
In commit 20272d460f, Monday, I blanked out on the syntax of groff
ms(7)'s font macros.  The opening bracket did not trail the `\c` once
the macro rearranged its arguments for formatting.

I had wondered why my new "stuff on the line after \c" diagnostics
didn't trip over this...it's because there was no such stuff after the
`I` macro did its work.

The superficially similar change to groff_ms(7) in the same commit _was_
a legitimate case, and, if reverted, is correctly flagged by the
formatter with a syntax warning.  It's better to have the formatter
catch these sorts of errors than to rely on human eyeballs, which is why
I implemented the warning.