[grotty]: Revise copyright notice in test script.
Identify myself as the copyright holder.  I've never executed copyright
assignment paperwork with the FSF, so my "legally significant"[1]
origination of this file in 2022 can't be under the FSF's copyright.

Drop ersatz '(C)' copyright symbol.  Software developers have long
labored under the no-longer-correct misconception that omitting a
copyright symbol from one's notice was a fatal defect that effectively
placed the work in the public domain.  That stopped being true as of 1
March 1989.[2]  Further, prior to guidance issued by the U.S. Copyright
Office in the decades since, the use of "(C)" as a substitute for a
copyright sign _may not have sufficed_ to prevent the copyright notice
from being regarded as defective.  The Copyright Office, then and now,
prefers the abbreviation "copr." when © is typographically
unavailable.[ibid.]  Nowadays, its advice is that "c" (note lowercase)
is an "acceptable variant", that _may_ retain the efficacy of the
copyright notice.  The word "copyright", spelled out in full, also
suffices per that resource, and is already present in this notice.

Drop stump comment after license information.

Update Vim modeline for human comprehension.

[1] "A change of just a few lines (less than 15 or so) is not legally
significant for copyright."  Conversely, >= 15 lines would be.

https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html

[2] https://www.copyright.gov/circs/circ03.pdf