[groff]: Revise copyright notices in two tests.
An analysis predicated on "a net increase of at least 15 lines in one
commit" results in a ratcheting back of the dates in the notice.

Credit TANAKA Takuji as contributor.

(I retain an FSF copyright notice corresponding to all changes not
authored by me or Ingo Schwarze--and which are above the "legal
significance" threshold--because I don't know who _has_ filed copyright
assignment paperwork with the FSF; I know only that Ingo and I haven't.)