src/devices/grops/psrm.cpp: Revise copyright.
Ratchet back ending date of notice to 2003(!), corresponding to a commit
authored by Werner Lemberg.

(I retain an FSF copyright notice corresponding to all changes not
authored by me or Ingo Schwarze--and which are above the "legal
significance" threshold[1]--because I don't know who _has_ filed
copyright assignment paperwork with the FSF; I know only that Ingo and I
haven't.)

$ git show --stat 87edb525555d70a53ca2754df7b26e9032fffe51 -- src/devices/grops/psrm.cc
commit 87edb525555d70a53ca2754df7b26e9032fffe51
Author: Werner LEMBERG <wl@gnu.org>
Date:   Mon Mar 10 21:51:55 2003 +0000
...
    Added support to access more than 256 glyphs in Type 1 fonts.
...
 src/devices/grops/psrm.cc | 107 +++++++++++++++++++++++++-------------
 1 file changed, 70 insertions(+), 37 deletions(-)

Every commit to this file since then has been below the GNU Maintainers'
Guide's threshold for "legal significance".

(At this point I must more clearly document an assumption that I am
making: that a "change" must constitute a net addition of "lines" to a
file.  If a change _deleted_ 15 or more lines, how does the
"originality" of copyright apply?  Similarly, if someone makes a
uniform syntactical or spelling change to a file, as one might do with
`sed 's/colour/color/', does that meet the threshold of
copyrightability, let alone the FSF's "legal[] significance", even if
it affects 200 lines of code or documentation?)

[1] "A change of just a few lines (less than 15 or so) is not legally
significant for copyright."  Conversely, >= 15 lines would be.

https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html