src/devices/grops/ps.h: Revise copyright.
Ratchet back ending date of notice to 1991(!!), corresponding to a
"commit" authored by James Clark, which is really just the initial
check-in of this file to our Git repository.

Every commit to this file since then has been below the GNU Maintainers'
Guide's threshold for "legal significance".

(At this point I must more clearly document an assumption that I am
making: that a "change" must constitute a net addition of "lines" to a
file.  If a change _deleted_ 15 or more lines, how does the
"originality" of copyright apply?  Similarly, if someone makes a
uniform syntactical or spelling change to a file, as one might do with
`sed 's/colour/color/', does that meet the threshold of
copyrightability, let alone the FSF's "legal[] significance", even if
it affects 200 lines of code or documentation?)

[1] "A change of just a few lines (less than 15 or so) is not legally
significant for copyright."  Conversely, >= 15 lines would be.

https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html