src/preproc/eqn/pile.cpp: Revise copyright notice.
Ratchet back ending date of notice to 2007, corresponding to a commit
authored by Werner Lemberg.

$ git show --stat d24ec56b925b63f9d89e01e087ab20a197537828 -- src/preproc/eqn/pile.cpp
commit d24ec56b925b63f9d89e01e087ab20a197537828
Author: Werner LEMBERG <wl@gnu.org>
Date:   Mon Feb 5 15:24:56 2007 +0000

...

 src/preproc/eqn/pile.cpp | 39 ++++++++++++++++++++++++++++-----------
 1 file changed, 28 insertions(+), 11 deletions(-)

(The commit log was not descriptive of the changes to this file, which
were mainly adjustments to code style.  One could argue about the
copyrightability or "legal significance" of those changes...with whoever
is responsible for the relevant material in the GNU Maintainers' Guide.)

Every commit to this file since then has been below the GNU Maintainers'
Guide's threshold for "legal significance".

[1] "A change of just a few lines (less than 15 or so) is not legally
significant for copyright."  Conversely, >= 15 lines would be.

https://www.gnu.org/prep/maintain/html_node/Legally-Significant.html