ChangeLog: Fix component annotations.
Our automated tests for the formatter program, troff, live in the
"tests" directory of the front end program, groff.  This is to give us
flexibility in how we prepare and test output, and to avoid the tedium
of setting up troff to run in the build tree; our test scripts delegate
that responsibility to the "test-groff" script.  It also makes every
test of the formatter an integration test as well, which has to date
proven advantageous without any exceptions I can think of.